Changed to uint64

+++ include/PetriEngine/ExplicitColored/SuccessorGenerator/ColoredSuccessorGenerator.h
@@ -26,7 +26,7 @@ namespace PetriEngine::ExplicitColored {
     public:
         explicit ColoredSuccessorGenerator(const ColoredPetriNet& net, Binding_t constraintThreshold);
         ~ColoredSuccessorGenerator() = default;
-        mutable uint32_t checkedBindings = 0;
+        mutable uint64_t checkedBindings = 0;
 
         std::pair<ColoredPetriNetStateFixed, TraceMapStep> next(ColoredPetriNetStateFixed& state) const {
             return _nextFixed(state);
